backend/requirements.txt
django==6.0
python-dotenv==1.0.1
djangorestframework
django-cors-headers
Pillow